Refactor v2 validator references
